requirements.txt
Django==1.5.1
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
south==0.8.4
django-floppyforms==1.1
djangorestframework==2.3.13